requirements.txt
Django==2.2.9
django-filter==2.2.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
Pillow==7.0.0
psycopg2==2.8.4
PyJWT==1.7.1
python-decouple==3.3
pytz==2019.3
sqlparse==0.3.0
